feat(levm): implement simulate_tx
