+++ crates/vm/backends/revm/mod.rs
@@ -1,10 +1,10 @@
 pub mod db;
 pub mod execution_db;
-pub mod execution_result;
 pub mod helpers;
 #[cfg(feature = "l2")]
 mod mods;
 
+use super::execution_result::ExecutionResult;
 use super::BlockExecutionResult;
 use crate::constants::{
     BEACON_ROOTS_ADDRESS, CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS, HISTORY_STORAGE_ADDRESS,
@@ -17,7 +17,6 @@ use ethrex_common::types::AccountInfo;
 use ethrex_common::{BigEndianHash, H256, U256};
 use ethrex_storage::{error::StoreError, AccountUpdate};
 
-use execution_result::ExecutionResult;
 use revm::db::states::bundle_state::BundleRetention;
 use revm::db::AccountStatus;
 use revm::{
